+++ .babelrc
@@ -1,0 +1,16 @@
+{ 
+  "presets": [
+    ["@babel/preset-env", {
+      "targets": {
+          "ie": "11"
+      }
+    }]
+  ],
+  "plugins": [
+    ["@babel/plugin-transform-runtime",
+      {
+        "regenerator": true
+      }
+    ]
+  ],
+}
