+++ babel.config.js
@@ -4,7 +4,7 @@ module.exports = {
       "@babel/preset-env",
       {
         targets: {
-          node: "current",
+          ie: "11",
         },
       },
     ],
