--- index.js
@@ -1,7 +1,0 @@
-const data = require("./data/fiches-travail.json");
-const { decode } = require("./src/email");
-
-module.exports = {
-  data,
-  decode,
-};
